--- src/import-service/src/handlers/importProductsFile.ts
@@ -1,23 +1,0 @@
-import { errorResponse, successResponse } from "../utils/apiResponseBuilder";
-
-export const importProductsFile =
-  ({ s3 }) =>
-  async (event) => {
-    try {
-      const catalogName = event.queryStringParameters.name;
-      const catalogPath = `uploaded/${catalogName}`;
-
-      const params = {
-        Bucket: "import-service-5",
-        Key: catalogPath,
-        Expires: 60,
-        ContentType: "text/csv",
-      };
-
-      const url = await s3.getSignedUrlPromise("putObject", params);
-
-      return successResponse(url);
-    } catch (error) {
-      errorResponse(error);
-    }
-  };
