--- src/import-service/src/handlers/index.ts
@@ -1,2 +1,0 @@
-export { importProductsFile } from './importProductsFile';
-export { importFileParser } from './importFileParser';
